// src/app/modules/project/project.interface.ts
export interface IProject {
  title: string;
  description: string;
  longDescription: string;
  image: string;
  category: string;
  tags: string[];
  demoUrl: string;
  githubUrl: string;
  challenge: string;
  solution: string;
  results: string[];
  duration: string;
  teamSize: string;
  client: string;
}